moved plugins from src to plugins folder
